requirements.txt
# Stock Prediction Application Requirements
# All dependencies - no hardcoded versions in code

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pmdarima>=2.0.0
scipy>=1.10.0
plotly>=5.14.0
python-dotenv>=1.0.0

